requirements.txt
pandas 
scikit-learn
torch torchvision
pytest
PyYAML